Use decimals for transitionDuration

+++ lib/src/models/ellipse.dart
@@ -18,7 +18,7 @@ class Ellipse extends Vector {
     LayoutAlign layoutAlign,
     LayoutConstraint constraints,
     String transitionNodeID,
-    int transitionDuration,
+    double transitionDuration,
     EasingType transitionEasing,
     double opacity,
     SizeRectangle absoluteBoundingBox,
